add removal, upload as a service

+++ mymodule.py
@@ -2,7 +2,21 @@
 # -*- coding: utf-8 -*-
 
 import os
+import os.path
 
 
-def rm(filename):
-    os.remove(filename)
+class RemovalService(object):
+    """A service for removing objects from the filesystem."""
+
+    @staticmethod
+    def rm(filename):
+        if os.path.isfile(filename):
+            os.remove(filename)
+
+
+class UploadService(object):
+    def __init__(self, removal_service):
+        self.removal_service = removal_service
+
+    def upload_complete(self, filename):
+        self.removal_service.rm(filename)
